requirements.txt
Flask
pandas
Authlib
requests
gspread
google-auth
